+++ classes/HrxData.php
@@ -111,9 +111,77 @@ class HrxData
         return $res;
     }
 
-    public static function getTerminalsByCountry($country_code, $item_list = null)
+    /**
+     * Retrieves all terminal locations for a given country code.
+     *
+     * This function calls `getLocationListByCountry()` to get each location in a memory-efficient manner.
+     * The results are then stored in an array and returned in a single array.
+     *
+     * @param string $country_code The country code to filter terminal locations by (e.g., "LT").
+     * @return array Returns an array of locations.
+     */
+    public static function getTerminalsByCountry($country_code)
     {
-        return HrxDeliveryTerminal::getLocationListByCountry($country_code);
+        $locations = array();
+        foreach (HrxDeliveryTerminal::getLocationListByCountry($country_code) as $location) {
+            $locations[] = $location;
+        }
+        return $locations;
+    }
+
+    public static function getCoordinatesByAddress($address, $country)
+    {
+        $url = 'https://geocode.arcgis.com/arcgis/rest/services/World/GeocodeServer/findAddressCandidates';
+        $query_array = array(
+            'f' => 'pjson',
+            'maxLocations' => 1,
+            'forStorage' => 'false',
+            'singleLine' => $address,
+            'sourceCountry' => $country,
+        );
+
+        $ch = curl_init();
+        curl_setopt($ch, CURLOPT_URL, $url . '?' . http_build_query($query_array));
+        curl_setopt($ch, CURLOPT_RETURNTRANSFER, 1);    
+        $responseJson = curl_exec($ch);
+        curl_close($ch);
+
+        $response = json_decode($responseJson, true);
+
+        if ( empty($response['candidates']) ) {
+            return false;
+        }
+
+        return array(
+            'latitude' => $response['candidates'][0]['location']['y'],
+            'longitude' => $response['candidates'][0]['location']['x'],
+        );
+    }
+
+    public static function calculateDistanceBetweenPoints($latitude_from, $longitude_from, $latitude_to, $longitude_to, $unit = 'km')
+    {
+        switch ($unit) {
+            case 'km': //kilometers
+                $earth_radius = 6371;
+            case 'mi': //miles
+                $earth_radius = 3959;
+                break;
+            default: //meters
+                $earth_radius = 6371000;
+                break;
+        }
+
+        $lat_from = deg2rad($latitude_from);
+        $lon_from = deg2rad($longitude_from);
+        $lat_to = deg2rad($latitude_to);
+        $lon_to = deg2rad($longitude_to);
+
+        $lat_delta = $lat_to - $lat_from;
+        $lon_delta = $lon_to - $lon_from;
+
+        $angle = 2 * asin(sqrt(pow(sin($lat_delta / 2), 2) + cos($lat_from) * cos($lat_to) * pow(sin($lon_delta / 2), 2)));
+
+        return $angle * $earth_radius;
     }
 
     /**